Move `ty::SymbolName` to its own little module (cute)
